Splitting into multiple Files
A new file for every bot-command the bot has to react to.
They are located in ./Modules/
Functions are migrated to their Files
