+++ Modules/Accept.js
@@ -1,0 +1,5 @@
+module.exports = {
+  start: function(Discord, client, msg) {
+
+  }
+};
